feat: add usage of authorization to controllers
- Protect endpoints with RolesAllowed annotation
- Add checks that user has rights to do remove actions

+++ web/src/main/kotlin/projects/controllers/ProjectController.kt
@@ -1,13 +1,19 @@
 package projects.controllers
 
+import exceptions.NotEnoughRightsException
+import org.eclipse.microprofile.jwt.Claim
+import org.eclipse.microprofile.jwt.Claims
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponses
 import projects.extensions.toProjectResponse
 import projects.models.ProjectCreationRequest
 import projects.models.ProjectResponse
 import projects.services.IProjectService
+import resumes.services.IResumeService
 import java.util.*
+import javax.annotation.security.RolesAllowed
 import javax.enterprise.context.RequestScoped
+import javax.inject.Inject
 import javax.ws.rs.DELETE
 import javax.ws.rs.GET
 import javax.ws.rs.POST
@@ -16,8 +22,17 @@ import javax.ws.rs.Path
 @RequestScoped
 @Path("projects")
 class ProjectController(
+    private var _resumeService: IResumeService,
     private var _projectService: IProjectService,
 ) {
+
+    @Inject
+    @Claim(standard = Claims.upn)
+    private lateinit var userId: String
+
+    @Inject
+    @Claim(standard = Claims.groups)
+    private lateinit var groups: Set<String>
 
     @APIResponses(
         APIResponse(responseCode = "200", description = "Project is created"),
@@ -25,6 +40,7 @@ class ProjectController(
         APIResponse(responseCode = "404", description = "Resume with sent id does not exist")
     )
     @POST
+    @RolesAllowed("User")
     fun add(projectCreationRequest: ProjectCreationRequest): ProjectResponse {
         val creationModel = projectCreationRequest.toCreationModel()
 
@@ -37,6 +53,7 @@ class ProjectController(
     )
     @GET
     @Path("/{id}")
+    @RolesAllowed("Admin")
     fun getById(id: UUID): ProjectResponse {
         val project = _projectService.getById(id)
 
@@ -48,6 +65,7 @@ class ProjectController(
     )
     @GET
     @Path("/user/{userId}")
+    @RolesAllowed("Admin")
     fun getAllByUserId(userId: UUID): List<ProjectResponse> {
         val projects = _projectService.getAllByUserId(userId)
 
@@ -60,7 +78,16 @@ class ProjectController(
     )
     @DELETE
     @Path("/{id}")
+    @RolesAllowed("User", "Admin")
     fun removeById(id: UUID): ProjectResponse {
+        val project = _projectService.getById(id)
+
+        val resume = _resumeService.getById(project.resumeId)
+
+        if (!groups.contains("Admin") && resume.userId != UUID.fromString(userId)) {
+            throw NotEnoughRightsException("You do not have access to this project!")
+        }
+
         return _projectService.removeById(id).toProjectResponse()
     }
 }
